+++ R/surv-roc_auc_survival.R
@@ -7,6 +7,11 @@
 #' @templateVar fn roc_auc_survival
 #' @template return-dynamic-survival
 #' @details
+#' ROC AUC survival is a metric that should be
+#' `r attr(roc_auc_survival, "direction")`d. The output ranges from
+#' `r metric_range(roc_auc_survival)[1]` to
+#' `r metric_range(roc_auc_survival)[2]`, with
+#' `r metric_optimal(roc_auc_survival)` indicating perfect discrimination.
 #'
 #' This formulation takes survival probability predictions at one or more
 #' specific _evaluation times_ and, for each time, computes the area under the
@@ -62,7 +67,8 @@ roc_auc_survival <- function(data, ...) {
 
 roc_auc_survival <- new_dynamic_survival_metric(
   roc_auc_survival,
-  direction = "maximize"
+  direction = "maximize",
+  range = c(0, 1)
 )
 
 #' @rdname roc_auc_survival