+++ R/surv-royston.R
@@ -6,6 +6,9 @@
 #' @templateVar fn royston_survival
 #' @template return
 #' @details
+#' Royston D statistic is a metric that should be
+#' `r attr(royston_survival, "direction")`d. The output ranges from 0 to 1, with
+#' higher values indicating better prognostic separation.
 #'
 #' Royston and Sauerbrei proposed $R^2_D$ as a measure of explained variation
 #' on the log relative hazard scale based on the authors’ D statistic.