+++ tests/testthat/test-num-rpd.R
@@ -79,3 +79,30 @@ test_that("na_rm argument check", {
     rpd_vec(1, 1, na_rm = "yes")
   )
 })
+
+test_that("range values are correct", {
+  direction <- metric_direction(rpd)
+  range <- metric_range(rpd)
+  perfect <- ifelse(direction == "minimize", range[1], range[2])
+  worst <- ifelse(direction == "minimize", range[2], range[1])
+
+  df <- tibble::tibble(
+    truth = c(5, 6, 2, 6, 4, 1, 3)
+  )
+
+  df$estimate <- df$truth
+  df$off <- df$truth + 1
+
+  expect_identical(
+    rpd_vec(df$truth, df$estimate),
+    perfect
+  )
+  if (direction == "minimize") {
+    expect_gt(rpd_vec(df$truth, df$off), perfect)
+    expect_lt(rpd_vec(df$truth, df$off), worst)
+  }
+  if (direction == "maximize") {
+    expect_lt(rpd_vec(df$truth, df$off), perfect)
+    expect_gt(rpd_vec(df$truth, df$off), worst)
+  }
+})