+++ tests/testthat/test-num-rpiq.R
@@ -71,3 +71,30 @@ test_that("na_rm argument check", {
     rpiq_vec(1, 1, na_rm = "yes")
   )
 })
+
+test_that("range values are correct", {
+  direction <- metric_direction(rpiq)
+  range <- metric_range(rpiq)
+  perfect <- ifelse(direction == "minimize", range[1], range[2])
+  worst <- ifelse(direction == "minimize", range[2], range[1])
+
+  df <- tibble::tibble(
+    truth = c(5, 6, 2, 6, 4, 1, 3)
+  )
+
+  df$estimate <- df$truth
+  df$off <- df$truth + 1
+
+  expect_identical(
+    rpiq_vec(df$truth, df$estimate),
+    perfect
+  )
+  if (direction == "minimize") {
+    expect_gt(rpiq_vec(df$truth, df$off), perfect)
+    expect_lt(rpiq_vec(df$truth, df$off), worst)
+  }
+  if (direction == "maximize") {
+    expect_lt(rpiq_vec(df$truth, df$off), perfect)
+    expect_gt(rpiq_vec(df$truth, df$off), worst)
+  }
+})